# api/openapi.yaml
openapi: 3.1.0
info:
  title: Distributed Job Scheduler API
  version: 1.0.0
  description: >
    Submit, schedule and manage background jobs. Conventions, error codes and roles are
    described in docs/low-level-design.md §9. Every request needs
    `Authorization: Bearer <api key>`.
servers:
  - url: http://localhost:8080
security:
  - apiKey: []
paths:
...
  /v1/schedules/{id}:
    parameters: [{ $ref: "#/components/parameters/ScheduleID" }]
    get:
      summary: Get a schedule, with its next five fire times
      responses:
        "200": { $ref: "#/components/responses/Schedule" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      summary: Update a schedule (operator); omitted fields are unchanged, null clears start_at, end_at, max_runs and priority
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ScheduleRequest" } } } }
      responses:
        "200": { $ref: "#/components/responses/Schedule" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      summary: Delete a schedule (operator); jobs that already ran stay in history
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    apiKey: { type: http, scheme: bearer, description: "jsk_<prefix>_<secret>" }
  parameters:
...
    ScheduleID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Error:
      description: Error (see LLD §9.2 for codes)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Schedule:
      description: The schedule
      content: { application/json: { schema: { $ref: "#/components/schemas/Schedule" } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, examples: [invalid_argument, not_found, conflict, rate_limited] }
            message: { type: string }
            request_id: { type: string }
            details: { type: object }
    Priority: { type: string, enum: [CRITICAL, HIGH, NORMAL, LOW] }
...
    Duration: { type: string, description: Go duration, examples: ["30s", "5m", "1h"] }
...
    Trigger:
      type: object
      required: [kind]
      description: "Cron (5 or 6 fields, or a macro such as @daily) in an IANA time zone, or a fixed-rate or fixed-delay interval (LLD §10.2)"
      properties:
        kind: { type: string, enum: [cron, fixed_rate, fixed_delay] }
        cron: { type: string, examples: ["0 2 * * *", "*/15 * * * *", "@hourly"] }
        time_zone: { type: string, default: UTC, examples: [Europe/Paris] }
        interval: { $ref: "#/components/schemas/Duration" }
    ScheduleRequest:
      type: object
      description: name, job_type and trigger are required on create; job_type cannot change
      properties:
        name: { type: string, pattern: "^[a-z0-9][a-z0-9._-]{0,99}$" }
        job_type: { type: string }
        payload: { description: "Any JSON value up to 64 KiB; defaults to {}" }
        labels: { type: object, additionalProperties: { type: string }, maxProperties: 16 }
        priority: { oneOf: [{ $ref: "#/components/schemas/Priority" }, { type: "null" }], description: "null uses the job type's default" }
        trigger: { $ref: "#/components/schemas/Trigger" }
        start_at: { type: [string, "null"], format: date-time }
        end_at: { type: [string, "null"], format: date-time }
        max_runs: { type: [integer, "null"], minimum: 1 }
        jitter: { $ref: "#/components/schemas/Duration" }
        misfire_policy: { type: string, enum: [fire_once, skip, fire_all], default: fire_once }
        overlap_policy: { type: string, enum: [skip, buffer_one, allow, cancel_previous], default: skip }
    Schedule:
      type: object
      description: "Jobs a schedule has materialized but that aren't due yet are provisional: pausing, editing or deleting the schedule withdraws them (ADR-016)."
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        job_type: { type: string }
        payload: {}
        labels: { type: object, additionalProperties: { type: string } }
        priority: { oneOf: [{ $ref: "#/components/schemas/Priority" }, { type: "null" }] }
        trigger: { $ref: "#/components/schemas/Trigger" }
        start_at: { type: [string, "null"], format: date-time }
        end_at: { type: [string, "null"], format: date-time }
        max_runs: { type: [integer, "null"] }
        jitter: { $ref: "#/components/schemas/Duration" }
        misfire_policy: { type: string }
        overlap_policy: { type: string }
        state: { type: string, enum: [ACTIVE, PAUSED, COMPLETED] }
        next_fire_at: { type: [string, "null"], format: date-time }
        last_fire_at: { type: [string, "null"], format: date-time }
        fire_count: { type: integer }
        upcoming: { type: array, items: { type: string, format: date-time }, description: Next five fire times (single reads only) }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...